requirements.txt
python-pptx>=1.0.0
PyYAML>=6.0
Flask>=3.0
gunicorn>=21.0
pytest>=8.0
